==> 01 Webscraping Wikipedia for Netflix original films.ipynb <==
# Netflix original films webscraping
### A webscraping script that reads in Wikipedia pages for lists of Netflix original films, turns them into pandas DataFrames, and then saves out as csv files.

import pandas as pd # for data analysis
import requests # need this to talk to a website
from bs4 import BeautifulSoup # to parse HTML documents
import numpy as np

## Let's generate a pandas DataFrame from tables in a Wikipedia page -- 2015-2017 for example

# check website response
wikiurl="https://en.wikipedia.org/wiki/List_of_Netflix_original_films_(2015-2017)"
table_class="wikitable sortable jquery-tablesorter"
response=requests.get(wikiurl)
print(response.status_code) # 200 means good to go

# parse html into a beautifulsoup object
soup = BeautifulSoup(response.text, 'html.parser') # get the content
table=soup.find_all('table',{'class':"wikitable"}) # find all <table> with "wikitable" class
print(len(table)) # 3 tables in this wikipedia page
print(table[0])

# find film types and add to individual table before concatenating
heading=soup.find_all('h2')

# these are common heading across wikipedia pages, so let's drop them
headings_to_drop = ['Contents','References','External links']

# loop over all headings and grab film type names
film_type = []
for h in range(len(heading)):
    name_str = str(heading[h])
    start_idx = name_str.find(">")+1 # where the name string starts
    end_idx = name_str.find("</") # where the name string ends
    heading_str = name_str[start_idx:end_idx]
    heading_str == headings_to_drop
    if not heading_str in headings_to_drop:
        film_type.append(heading_str)

# make sure that the number of film types match the number of tables we grabbed earlier
if len(table) != len(film_type):
    print("Warning: the number of film types does not match the number of tables")

print(film_type)


# now concatenate all three tables while adding the film type as a new column at the same time
dfs = []
for d in range(len(table)):
    df=pd.read_html(str(table[d]))
    # convert list to dataframe
    df=pd.DataFrame(df[0])
    df['Film type'] = film_type[d]
    dfs.append(df)
final_df = pd.concat(dfs, ignore_index=True)
print(final_df)
# final_df.describe()

## Do the same for all other pages -- writing a function that we can pass in the url to make this easier

def wikitable2df(url,add_heading=False,heading_name=None):
    '''
    reads in tables in a wikipedia page and converts into a pandas DataFrame.
    '''
    from io import StringIO # for pd.read_html
    
    # check website response
    table_class="wikitable sortable jquery-tablesorter"
    response=requests.get(url)
    if response.status_code != 200:
        raise ValueError("No-go on the status_code: May not be allowed to web-scrape from this website")

    # parse html into a beautifulsoup object
    soup = BeautifulSoup(response.text, 'html.parser') # get the content
    table=soup.find_all('table',{'class':"wikitable"}) # find all <table> with "wikitable" class

    # if add_heading = True, grab heading values and add an additional column filled with that value
    if add_heading:
        # find headings and add to individual table before concatenating
        heading=soup.find_all('h2')
        
        # these are some common headings that can exist in any wikipedia pages, so let's drop them
        headings_to_drop = ['Contents','References','External links','Notes']
        
        # loop over all headings and grab heading strings
        new_col = []
        for h in range(len(heading)):
            name_str = str(heading[h])
            start_idx = name_str.find(">")+1 # where the name string starts
            end_idx = name_str.find("</") # where the name string ends
            heading_str = name_str[start_idx:end_idx]
            if not heading_str in headings_to_drop:
                new_col.append(heading_str)
        
        # make sure that the number of headings match the number of tables we grabbed earlier
        if len(table) != len(new_col):
            print(url, new_col)
            raise ValueError("The number of headings does not match the number of tables")
        
    # now concatenate all tables while adding the film type as a new column at the same time
    dfs = []
    for d in range(len(table)):
        df=pd.read_html(StringIO(str(table[d])))
        # convert list to dataframe
        df=pd.DataFrame(df[0])
        df[str(heading_name)] = new_col[d]
        dfs.append(df)
    final_df = pd.concat(dfs, ignore_index=True)
    return final_df

Now we can run this function over all urls and then concatenate the returned DataFrames.

urls = ["https://en.wikipedia.org/wiki/List_of_Netflix_original_films_(2015-2017)",
        "https://en.wikipedia.org/wiki/List_of_Netflix_original_films_(2018)",
        "https://en.wikipedia.org/wiki/List_of_Netflix_original_films_(2019)",
        "https://en.wikipedia.org/wiki/List_of_Netflix_original_films_(2020)",
        "https://en.wikipedia.org/wiki/List_of_Netflix_original_films_(2021)",
        "https://en.wikipedia.org/wiki/List_of_Netflix_original_films_(2022)",
        "https://en.wikipedia.org/wiki/List_of_Netflix_original_films_(2023)",
        "https://en.wikipedia.org/wiki/List_of_Netflix_original_films_(2024)"]
        # "https://en.wikipedia.org/wiki/List_of_Netflix_original_films_(since_2025)"] excluding 2025- page for now
        
dfs = []
for url in urls:
    df = wikitable2df(url,add_heading=True,heading_name='Film type')
    dfs.append(df)
final_df = pd.concat(dfs, ignore_index=True)
final_df

Let's clean the data a little bit before saving

# Add index as a column
final_df['id'] = final_df.index
final_df_replaced = final_df.replace('', np.nan)  # Assuming empty cells are represented as empty strings
final_df.info()

final_df_replaced.info()

final_df_replaced[final_df_replaced['Title']=='My Own Man']

Let's save out as a csv file.

final_df_replaced.to_csv('netflix_og_films.csv',encoding='utf-8-sig',index=False) # using utf-8-sig for special characters in the title
