==> 02 Creating a Database with IMDB and Netflix Data.ipynb <==
# Install dependencies as needed:
# !pip install --upgrade kagglehub
import kagglehub

# Download latest version
path = kagglehub.dataset_download("ashirwadsangwan/imdb-dataset/version/824") # if already downloaded, will still give us the path

print("Path to dataset files:", path)

# check if the files are successfully downloaded in the said path
!ls -alh $path

## We'll use two files that contain basic information and ratings for each title.
Note: if the title matching is not successful with the titles in the basics file, try akas file where there is localized/original titles and language information which may help with matching.

### title.basics.tsv.gz - Contains the following information for titles:
- tconst (string) - alphanumeric unique identifier of the title.
- titleType (string) – the type/format of the title (e.g. movie, short,
tvseries, tvepisode, video, etc).
- primaryTitle (string) – the more popular title / the title used by
the filmmakers on promotional materials at the point of release.
- originalTitle (string) - original title, in the original language.
- isAdult (boolean) - 0: non-adult title; 1: adult title.
- startYear (YYYY) – represents the release year of a title. In the
case of TV Series, it is the series start year.
- endYear (YYYY) – TV Series end year. for all other title types.
- runtimeMinutes – primary runtime of the title, in minutes.
- genres (string array) – includes up to three genres associated with
the title.

### title.ratings.tsv.gz – Contains the IMDb rating and votes information for titles:
- tconst (string) - alphanumeric unique identifier of the title.
- averageRating – weighted average of all the individual user ratings.
- numVotes - number of votes the title has received.

Load the two tsv files to dataframes and then write to a sqlite table.

import pandas as pd
import os
import sqlite3 

# Connect to SQLite database 
conn = sqlite3.connect('films.db') 

# Create a cursor object 
cur = conn.cursor() 

# Load TSV file into a DataFrame, specifying tab as the separator
df_basics = pd.read_csv(os.path.join(path,'title.basics.tsv'), sep='\t')
df_basics = df_basics.convert_dtypes()
df_basics.info()

# Change data types to correct types
df_basics['isAdult'] = df_basics['isAdult'].astype(bool)
df_basics['startYear']=pd.to_numeric(df_basics['startYear'], errors='coerce')
df_basics['endYear']=pd.to_numeric(df_basics['endYear'], errors='coerce')
df_basics['runtimeMinutes']=pd.to_numeric(df_basics['runtimeMinutes'], errors='coerce')

# Drop rows where titleType is 'tvEpisode' to reduce memory usage since we're only looking at movies
df_basics = df_basics[df_basics['titleType']!='tvEpisode']

# Check before saving to database
df_basics.info()

# Do the same with the ratings table
# Load TSV file into a DataFrame, specifying tab as the separator
df_ratings = pd.read_csv(os.path.join(path,'title.ratings.tsv'), sep='\t')
df_ratings = df_ratings.convert_dtypes()
df_ratings.info()

# Write the data to a sqlite table 
df_basics.to_sql('imdb_basics', conn, if_exists='replace', index=False) 
df_ratings.to_sql('imdb_ratings', conn, if_exists='replace', index=False) 

# Check if basics dataframe was saved correctly by loading and pulling up first 10 entries
query = """
        SELECT *
        FROM imdb_basics
        LIMIT 10;
        """
for row in cur.execute(query): 
    print(row) 

# Check if ratings dataframe was saved correctly by loading and pulling up first 10 entries
query = """
        SELECT *
        FROM imdb_ratings
        LIMIT 10;
        """
for row in cur.execute(query): 
    print(row) 

### Let's load Netflix original film list and add to the database

First, I want to add the netflix data to the database so we can access everything from one place in the future.

# Load TSV file into a DataFrame, specifying tab as the separator
df_netflix = pd.read_csv('netflix_og_films.csv')
# Fix the column names before writing to the database to follow SQL syntax
df_netflix.rename(columns={'Title': 'title', 'Release date': 'releaseDate', 
                   'Genre':'genre','Runtime':'runtime','Language':'language','Film type':'filmType'}, inplace=True)

Before we save this dataframe into the database, I want to convert runtime ('x h x min') to minutes, so that we can use this as an additional constraint when matching the title of the films to IMDB.

def hm2m(time_str):
    """a function to convert runtime to minutes"""
    cuml_min = 0
    if time_str.find('h')>0:
        h_pos = time_str.find('h')
        cuml_min += int(time_str[:h_pos-1])*60
        time_str = time_str[h_pos+2:]
    
    if time_str.find('min')>0:
        m_pos = time_str.find('min')
        cuml_min += int(time_str[:m_pos-1])
        
    return cuml_min

# Convert runtime that's in X h X min to minutes
df_netflix = df_netflix.convert_dtypes() # convert to string first
df_netflix['runtime_min']=df_netflix['runtime'].apply(hm2m) # apply the above function to the runtime column
df_netflix.head() # check if it worked correctly

df_netflix['releaseYear'] = df_netflix['releaseDate'].str[-4:].astype(int)

df_netflix.head()

df_netflix.info()

# Write the data to a sqlite table 
df_netflix.to_sql('netflix', conn, if_exists='replace', index=False) 

# Read in data and list the column names and the first 10 entries
query = """
        SELECT *
        FROM netflix
        LIMIT 10;
        """
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)
for row in out: 
    print(row) 

Close out the connection to SQLite database

# Close connection to SQLite database 
conn.close()
