==> 04 Adding More Rating Data to the Database.ipynb <==
### Download and add Rotten Tomatoes dataset to the database

# Install dependencies as needed:
# !pip install --upgrade kagglehub
import kagglehub

# Download latest version
path = kagglehub.dataset_download("andrezaza/clapper-massive-rotten-tomatoes-movies-and-reviews/versions/4/rotten_tomatoes_movies.csv",
                                  path="rotten_tomatoes_movies.csv",force_download=True)

print("Path to dataset files:", path)

# check if the files are successfully downloaded in the said path
# on Mac
!ls -alh $path
# on Windows
!dir $path 

## Using the _movies.csv file that contains basic information and ratings for each movie.

### rotten_tomatoes_movies.csv - Each record represents a movie available on Rotten Tomatoes and includes all fields available on the website
*Note that 'rating' field here refers to the age-based rating (e.g. G) and not the rating by viewers.

- id - Unique identifier for each movie
- title - The title of the movie
- audienceScore - The average score given by regular viewers (0-100)
- tomatoMeter - The percentage of positive reviews from professional critics (0-100)
- rating - The movie's age-based classification (e.g., G, PG, PG-13, R)
- ratingContents - Content leading to the rating classification
- releaseDateTheaters - The date the movie was released in theaters
- releaseDateStreaming - The date the movie became available for streaming
- runtimeMinutes - The duration of the movie in minutes
- genre - The movie's genre(s)
- originalLanguage - The original language of the movie
- director - The movie's director
- writer - The writer(s) responsible for the movie's screenplay
- boxOffice - The movie's total box office revenue
- distributer - The company responsible for distributing the movie
- soundMix - The audio format(s) used in the movie

Load the two tsv files to dataframes and then write to a sqlite table.

import pandas as pd
import os
import sqlite3 
import numpy as np

# Load CSV file into a DataFrame
df_rt = pd.read_csv(path)
df_rt = df_rt.convert_dtypes()
df_rt.info()

df_rt.head()

Cleaning the data and adding some more informative columns that can be used to match entries between datasets.
1. add releaseYear column from releaseDateTheaters and releaseDateStreaming, use the earlier one if both exists and do not match.
2. remove entries that don't have both audienceScore and tomatoMeter

# 1. Convert release date to release year and add as a new column
# loading datetime package to make extracting year easy
import datetime

# First converting releaseDateStreaming to release year
# Set values that don't follow the 'YYYY-MM-DD' format to nan
tmp = np.where(df_rt['releaseDateStreaming'].fillna('nan').str.len() == 10, df_rt['releaseDateStreaming'],np.nan)
# Convert values to datetime data type
tmp = pd.to_datetime(tmp,format='ISO8601')
# Extract year from datetime object
tmp = tmp.year
# Convert year values to integers - swap NA with 10000 - and add as a new column
df_rt['releaseYearStreaming'] = tmp.fillna(10000).astype(int)
# Compare the original column and the new column
df_rt[['releaseDateStreaming','releaseYearStreaming']].head()

# Now doing the same for releaseDateTheaters
# Set values that don't follow the 'YYYY-MM-DD' format to nan
tmp = np.where(df_rt['releaseDateTheaters'].fillna('nan').str.len() == 10, df_rt['releaseDateTheaters'],np.nan)
# Convert values to datetime data type
tmp = pd.to_datetime(tmp,format='ISO8601')
# Extract year from datetime object
tmp = tmp.year
# Convert year values to integers - swap NA with 10000 - and add as a new column
df_rt['releaseYearTheaters'] = tmp.fillna(10000).astype(int)
# Compare the original column and the new column
df_rt[['releaseDateTheaters','releaseYearTheaters']]

# Grab the lower value of the two - if both are NA, the value will be 10000
tmp_min = np.where(df_rt['releaseYearStreaming']>df_rt['releaseYearTheaters'],df_rt['releaseYearTheaters'],df_rt['releaseYearStreaming'])
df_rt['releaseYearEarlier'] = np.where(tmp_min==10000,np.nan,tmp_min)
# df_rt['releaseYearEarlier'] = df_rt['releaseYearEarlier'].astype('int64')
df_rt[['releaseYearStreaming','releaseYearTheaters','releaseYearEarlier']]

df_rt.info()

# 2. remove entries that don't have both audienceScore and tomatoMeter
df_rt=df_rt.dropna(axis='index',how='all',subset=['audienceScore','tomatoMeter'])
df_rt.head()

# Drop the temporary columns created
df_rt = df_rt.drop(columns = ['releaseYearStreaming','releaseYearTheaters'])

# Check dataframe for the last time before saving to the database
df_rt.head()

# Connect to SQLite database 
conn = sqlite3.connect('films.db') 

# Create a cursor object 
cur = conn.cursor() 

# Write the data to a sqlite table 
df_rt.to_sql('rotten_tomatoes', conn, if_exists='replace', index=False) 

# Check if the dataframe was saved correctly by loading and pulling up first 10 entries
query = """
        SELECT *
        FROM rotten_tomatoes
        LIMIT 10;
        """
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)

for row in cur.execute(query): 
    print(row) 

# Close connection to SQLite database 
conn.close() 

### Download and add Letterboxd dataset to the database

# Download latest version
path = kagglehub.dataset_download("gsimonx37/letterboxd", path = "movies.csv")

print("Path to dataset files:", path)

unzip(path)

### movies.csv
- id - movie identifier (primary key)
- name - the name of the film
- date - year of release of the film
- tagline - the slogan of the film
- description - description of the film
- minute - movie duration (in minutes)
- rating - rating (0-5 scale)

print(path)
!dir $path

# Load CSV file into a DataFrame
df_lb = pd.read_csv(path)
df_lb = df_lb.convert_dtypes()
df_lb.info()

df_lb.head()

Not much to adjust except resetting the id to string and dropping rows with NA rating.

df_lb['id']=df_lb['id'].astype(str)
df_lb.info()

df_lb=df_lb.dropna(axis='index',subset='rating')
df_lb.info()

# Drop the temporary columns created
df_lb = df_lb.drop(columns = ['tagline','description'])

# Connect to SQLite database 
conn = sqlite3.connect('films.db') 

# Create a cursor object 
cur = conn.cursor() 

# Write the data to a sqlite table 
df_lb.to_sql('letterboxd', conn, if_exists='replace', index=False) 

# Check if the dataframe was saved correctly by loading and pulling up first 10 entries
query = """
        SELECT *
        FROM letterboxd
        LIMIT 10;
        """
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)

for row in cur.execute(query): 
    print(row) 

# Close connection to SQLite database 
conn.close()
