==> 05 Combining Additional Data to Netflix Data.ipynb <==
### Before combining all data, I want to check how many matches to the Netflix data we can get with the Rotten Tomatoes and Letterboxd data, separately.

Below are the fields of the tables saved in the database.

### netflix – Contains the Netflix original film list (webscraped from Wikipedia):
- title (string) - official Netflix title
- releaseDate (string) - release date in '{month} {day}, {year}' format (e.g. 'October 16, 2015') 
- genre - main genre
- runtime - runtime of the flim in '{hour} h {minute} min' format (e.g. '1 h 40 min')
- language - main langauge of the film
- filmType - type of the film (e.g. feature film, documentary, special, etc)
- runtime_min - runtime converted to minutes

### rotten_tomatoes
- id (string) - Unique identifier for each movie
- title - The title of the movie
- audienceScore - The average score given by regular viewers (0-100)
- tomatoMeter - The percentage of positive reviews from professional critics (0-100)
- rating - The movie's age-based classification (e.g., G, PG, PG-13, R)
- ratingContents - Content leading to the rating classification
- releaseDateTheaters - The date the movie was released in theaters
- releaseDateStreaming - The date the movie became available for streaming
- runtimeMinutes - The duration of the movie in minutes
- genre - The movie's genre(s)
- originalLanguage - The original language of the movie
- director - The movie's director
- writer - The writer(s) responsible for the movie's screenplay
- boxOffice - The movie's total box office revenue
- distributer - The company responsible for distributing the movie
- soundMix - The audio format(s) used in the movie
- releaseYearEarlier - The year the movie was released, either in theaters or on a streaming service, whichever is earlier

### letterboxd
- id - unique identifier for each movie
- name - title of the movie
- date - release year of the movie
- minute - runtime in minutes
- rating - rating on letterboxd

import pandas as pd
import os
import sqlite3 

# Connect to SQLite database 
conn = sqlite3.connect('films.db') 

# Create a cursor object 
cur = conn.cursor() 

# Check if rotten_tomatoes dataframe was saved correctly by loading and pulling up first 10 entries
query = """
        SELECT *
        FROM rotten_tomatoes
        LIMIT 10;
        """
# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)
for row in out: 
    print(row) 

Let's try joining netflix and rotten_tomatoes table on the title. Starting with the least constraint to see how many matches just based on title we can get.

# Joining on title, year, and language and checking how many non-unique entries we get
query = """
        SELECT *
        FROM netflix AS n
        INNER JOIN rotten_tomatoes AS r
        ON LOWER(n.title) = LOWER(r.title) AND n.releaseYear = r.releaseYearEarlier AND n.language = r.originalLanguage
        GROUP BY n.id
        HAVING COUNT(*) > 1;
        """

# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)

for row in out: 
    print(row) 

# checking one entry from the output above in the rotten_tomatoes table
query = """
        SELECT *
        FROM rotten_tomatoes
        WHERE id = 'the_highwaymen';
        """

# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)

for row in out: 
    print(row) 

Seems like there are entries with the same id in rotten_tomatoes. Going to group by the id first and then join to prevent duplicates. Adding in the runtime constraint to make sure we are getting "close to" precise matches as much as possible.

query = """
        SELECT COUNT( DISTINCT n.id ) AS unique_id, COUNT(*) AS total
        FROM netflix AS n
        INNER JOIN (
            SELECT *
            FROM rotten_tomatoes 
            GROUP BY id) AS r
        ON LOWER(n.title) = LOWER(r.title) AND n.releaseYear = r.releaseYearEarlier AND n.language = r.originalLanguage
        WHERE ( r.runtimeMinutes <= n.runtime_min +5 AND r.runtimeMinutes >= n.runtime_min -5 );
        """

# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)

for row in out: 
    print(row) 

Moving onto letterboxd.

# Check if letterboxd dataframe was saved correctly by loading and pulling up first 10 entries
query = """
        SELECT *
        FROM letterboxd
        LIMIT 10;
        """
# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)
for row in out: 
    print(row) 

Let's proactively check if there are entries with duplicate ids in the letterboxd table as in the rotten_tomatoes table.

query = """
        SELECT id, COUNT(id)
        FROM letterboxd
        GROUP BY id
        HAVING COUNT(id)>1;
        """
# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)
for row in out: 
    print(row) 

Seems like all entries have unique ids! Moving onto combining with netflix data.

# Joining on title/name and year/date, and checking for non-unique entries
query = """
        SELECT *
        FROM netflix AS n
        INNER JOIN letterboxd AS l
        ON LOWER(n.title) = LOWER(l.name) AND n.releaseYear = l.date
        GROUP BY n.id
        HAVING COUNT(*) > 1;
        """

# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)

for row in out: 
    print(row) 

Let's add in the runtime constraint.

# Adding runtime constraint and checking the number of unique entries
query = """
        SELECT COUNT( DISTINCT n.id ) AS unique_id, COUNT(*) AS total
        FROM netflix AS n
        INNER JOIN letterboxd AS l
        ON LOWER(n.title) = LOWER(l.name) AND n.releaseYear = l.date
        WHERE ( l.minute <= n.runtime_min +5 AND l.minute >= n.runtime_min -5 );
        """

# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)

for row in out: 
    print(row) 

# Check the entry in netflix and letterboxd
query = """
        SELECT *
        FROM netflix
        WHERE title = 'Bubble';
        """
# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)
for row in out: 
    print(row) 
    
query = """
        SELECT *
        FROM letterboxd
        WHERE name = 'Bubble';
        """
# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)
for row in out: 
    print(row) 

TODO: maybe would be good to include primary language in letterboxd table and use as a condition to join as well?

# Load CSV file into a DataFrame\
path = '/Users/sunyoungpark/.cache/kagglehub/datasets/gsimonx37/letterboxd/versions/2/languages.csv'
df = pd.read_csv(path)
df = df.convert_dtypes()
df.info()

# Write the data to a sqlite table 
df.to_sql('letterboxd_languages', conn, if_exists='replace', index=False) 

query = """
        SELECT l.id, l.name, l.date, l.minute, l.rating, lang.language
        FROM letterboxd AS l
        LEFT JOIN letterboxd_languages AS lang
        ON l.id = lang.id
        WHERE lang.type = 'Primary language'
        LIMIT 10;
        """
# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)
for row in out: 
    print(row) 

Now using this joined table as a subquery to join to the netflix table

# Adding language match as another join condition and checking the number of unique entries
query = """
        SELECT COUNT( DISTINCT n.id ) AS unique_id, COUNT(*) AS total
        FROM netflix AS n
        INNER JOIN (
            SELECT l.id, l.name, l.date, l.minute, l.rating, lang.language
            FROM letterboxd AS l
            LEFT JOIN letterboxd_languages AS lang
            ON l.id = lang.id
            WHERE lang.type = 'Primary language'
            ) AS l
        ON LOWER(n.title) = LOWER(l.name) AND n.releaseYear = l.date AND n.language = l.language;
        """
# WHERE ( l.minute <= n.runtime_min +5 AND l.minute >= n.runtime_min -5 )
# print column names
out = cur.execute(query)
names = list(map(lambda x: x[0], cur.description))
print(names)

for row in out: 
    print(row) 

# Close connection to SQLite database 
conn.close()
